Split loudspeaker specs from speaker instance
Loudspeakers of a certain type are located somewhere in a room.
Previously, our class representing Loudspeakers encoded speaker
specifications and accepted position and axis normal as arguments to
most methods.

It is more natural to express speaker specifications (which are
typically identical for all speaker instances in the room) in one class
that can be composed within a class for each instance that expresses
location data, etc.