Speaker: make left and right speakers mirror
Previously, the same acoustic center was used for both the left and right
speaker. This was fine when the acoustic center was in the center of the
speaker in the y-axis. For speakers where the acoustic center is offset
from the center of the mesh. the speakers must be mirror images of each
other in the y-axis.